fix: moved types to models file
